ai-service/requirements.txt
Flask
Flask-Cors
ultralytics
sentence-transformers
scikit-learn
opencv-python-headless
Pillow
numpy
